requirements.txt
# Core Dependencies
streamlit==1.48.0
langchain==0.3.27
langchain-core==0.3.74
langgraph==0.6.5
# langsmith==0.2.11
langsmith==0.3.45

# LLM and Embeddings
mistralai==1.2.0

# Search and Web Scraping  
brave-search-python-client==0.4.27
beautifulsoup4==4.13.4
requests==2.32.3
lxml==5.3.0

# Vector Storage
faiss-cpu==1.9.0
numpy==2.1.0

# Utilities
python-dotenv==1.0.1
pydantic==2.11.0
typing-extensions==4.12.2

# Optional Dependencies
pandas==2.2.2
